requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# PDF processing
PyPDF2>=3.0.1

# Text processing and embeddings
langchain-text-splitters>=0.3.0
sentence-transformers>=3.0.0

# Vector database
chromadb>=0.5.0

# Keyword search
rank-bm25>=0.2.2

# LLM APIs
anthropic>=0.34.0

# Web interface
streamlit>=1.35.0

# Scientific computing
numpy>=1.26.0

# Utilities
tqdm>=4.66.1


google-genai
python-dotenv